Fix resize event listener duplication
Due to the way the resize event listener was written, a new one was
created on each render of the App component. This has been fixed
by adding the listener in a useEffect method that only fires once
the component is mounted. Because of what component this is, there
is no need for a remove clean up as it will never unmount.